# Conflicts:
#	application.properties
#	pom.xml
#	src/main/java/project/models/Person.java

+++ src/main/java/project/repositories/PostRepository.java
@@ -5,6 +5,5 @@ import org.springframework.stereotype.Repository;
 import project.models.Post;
 
 @Repository
-public interface PostRepository extends CrudRepository<Post, Long> {
-
+public interface PostRepository extends CrudRepository<Post, Integer> {
 }
